przerobienie programu z zapisywania i obsługi pliku na zapisywanie i obsługę bazy danych cd
usunięto klasy:
    - DelegationAutoincrementIdService
    - DelegationAddToFileService
    - DelegationsLoadFromFile
    - DelegationAcceptDiscardSaveToFile

+++ web-app/src/main/java/com/isa/cm3/dao/DelegationDao.java
@@ -15,7 +15,7 @@ public class DelegationDao {
     @PersistenceContext
     private EntityManager entityManager;
 
-    public Integer save(Delegation d) {
+    public Long save(Delegation d) {
         entityManager.persist(d);
         return d.getId();
     }
